refactor(everything): Refactor everything so that it uses itself. lerna-semantic-release now uses le
