+++ packages/lerna-semantic-release-io/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  'default': require('./io/index'),
+  mocks: require('./mocks/io/index'),
+  makeMockTask: require('./mocks/make-mock-task')
+};
