+++ packages/lerna-semantic-release-io/io/index.js
@@ -1,0 +1,15 @@
+var fs = require('./fs');
+var git = require('./git');
+var lerna = require('./lerna');
+var npm = require('./npm');
+var semanticRelease = require('./semantic-release');
+var shell = require('./shell');
+
+module.exports = {
+  fs: fs,
+  npm: npm,
+  git: git,
+  lerna: lerna,
+  shell: shell,
+  semanticRelease: semanticRelease
+};
