+++ packages/lerna-semantic-release-io/io/npm.js
@@ -1,0 +1,17 @@
+var execAsTask = require('lerna-semantic-release-utils').execAsTask;
+
+module.exports = {
+  publish: function version (path) {
+    return execAsTask('npm publish ' + path);
+  },
+  version: function version (v) {
+    return execAsTask('npm version ' + v + ' --git-tag-version false');
+  },
+  getVersion: function getVersion (packageName) {
+    return function (done) {
+      execAsTask(['npm view', packageName, 'version'].join(' '))(function (err, stdout) {
+        done(err, stdout.trim());
+      });
+    }
+  }
+};
