+++ packages/lerna-semantic-release-io/mocks/make-mock-task.js
