+++ packages/lerna-semantic-release-pre/package.json
@@ -1,0 +1,41 @@
+{
+  "name": "lerna-semantic-release-pre",
+  "author": "Joshua Nelson <jonelson@atlassian.com>, Joscha Feth <jfeth@atlassian.com>",
+  "bugs": {
+    "url": "https://github.com/atlassian/lerna-semantic-release/issues"
+  },
+  "dependencies": {
+    "lerna-semantic-release-utils": "^0.0.0",
+    "lerna-semantic-release-io": "^0.0.0",
+    "async": "~1.5.0",
+    "npmconf": "~2.1.0",
+    "rc": "~1.1.0",
+    "semantic-release": "~4.3.0"
+  },
+  "devDependencies": {
+    "mocha": "~2.5.0",
+    "expect.js": "~0.3.0",
+    "chai": "~3.5.0",
+    "sinon": "~1.17.0"
+  },
+  "engines": {
+    "node": ">=4",
+    "npm": ">=3"
+  },
+  "license": "MIT",
+  "main": "index.js",
+  "files": [
+    "bin"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/atlassian/lerna-semantic-release.git"
+  },
+  "scripts": {
+    "test": "mocha test/unit.js"
+  },
+  "release": {
+    "branch": "caribou"
+  },
+  "version": "0.0.0"
+}
