+++ packages/lerna-semantic-release-utils/index.js
@@ -1,0 +1,7 @@
+module.exports = {
+  bindTasks: require('./bind-tasks'),
+  execAsTask: require('./exec-as-task'),
+  forEachPackage: require('./for-each-package'),
+  log: require('./log'),
+  tagging: require('./tagging')
+};
