+++ packages/lerna-semantic-release-utils/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "lerna-semantic-release-utils",
+  "author": "Joshua Nelson <jonelson@atlassian.com>, Joscha Feth <jfeth@atlassian.com>",
+  "bugs": {
+    "url": "https://github.com/atlassian/lerna-semantic-release/issues"
+  },
+  "dependencies": {
+    "shelljs": "^0.7.0",
+    "winston": "^2.2.0",
+    "async": "^1.5.0"
+  },
+  "engines": {
+    "node": ">=4",
+    "npm": ">=3"
+  },
+  "license": "MIT",
+  "main": "index.js",
+  "files": [
+    "bin"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/atlassian/lerna-semantic-release.git"
+  },
+  "release": {
+    "branch": "caribou"
+  },
+  "scripts": {
+    "test": "echo 'no tests'; exit 0"
+  },
+  "version": "0.0.0"
+}
