+++ packages/lerna-semantic-release/bin/lerna-semantic-release.js
@@ -1,0 +1,24 @@
+#!/usr/bin/env node
+
+var task = process.argv[2];
+var io = require('lerna-semantic-release-io');
+
+var tasks = {
+  pre: require('lerna-semantic-release-pre'),
+  perform: require('lerna-semantic-release-perform'),
+  post: require('lerna-semantic-release-post')
+};
+
+function erorrHandler(err) {
+  console.error(err);
+  process.exit(+!!err);
+}
+
+try {
+  tasks[task]({
+    io: io,
+    callback: erorrHandler
+  });
+} catch(err) {
+  erorrHandler(err);
+}
