--- src/io.js
@@ -1,15 +1,0 @@
-var fs = require('../src/io/fs');
-var git = require('../src/io/git');
-var lerna = require('../src/io/lerna');
-var npm = require('../src/io/npm');
-var semanticRelease = require('../src/io/semantic-release');
-var shell = require('../src/io/shell');
-
-module.exports = {
-  fs: fs,
-  npm: npm,
-  git: git,
-  lerna: lerna,
-  shell: shell,
-  semanticRelease: semanticRelease
-};
