--- src/io/git.js
@@ -1,30 +1,0 @@
-var gitHead = require('git-head');
-var execAsTask = require('../utils/exec-as-task');
-var simpleGit = require('simple-git');
-
-module.exports = {
-  tag: function commit (tag, message) {
-    return execAsTask('git tag -am"' + message + '" ' + tag)
-  },
-  tagDelete: function tagDelete (tags) {
-    return execAsTask('git tag -d ' + tags.join(' '))
-  },
-  tagList: function tagList () {
-    return function (done) {
-      simpleGit().tags(done)
-    };
-  },
-  commit: function commit (message) {
-    return execAsTask('git commit -anm\'' + message +'\' --allow-empty');
-  },
-  head: gitHead,
-  push: function push () {
-    return execAsTask('git push origin');
-  },
-  pushTags: function pushTags () {
-    return execAsTask('git push origin --tags');
-  },
-  add: function add (file) {
-    return execAsTask('git add ' + file);
-  }
-};
