--- test/unit.js
@@ -1,3 +1,0 @@
-require('./steps/pre.js');
-require('./steps/perform.js');
-require('./steps/post.js');
